pyproject.toml
[project]
name = "dharma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
 "zenml==0.93.1",
 "kubernetes>=34.1.0",
]

[project.scripts]
forge = "clap.__main__:app"

[tool.uv]
package = true


[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
data-collection = [
    "minio>=7.2.20",
    "boto3>=1.38.27",
    "datasets>=3.6.0",
    "pillow>=11.2.1",
    "s3fs>=0.4.2",
    "slack-sdk>=3.35.0",
    "openai>=2.17.0",
    "tenacity>=9.1.4",
    "pdf2image>=1.17.0",
]
lint = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
]
infrastructure = [
    "typer>=0.21.0",
    "rich",
    "pulumi>=3.181.0",
    "pulumi-kubernetes>=4.23.0",
    "pulumi-command>=1.1.0",
    "python-dotenv>=1.1.1",
    "infisicalsdk>=1.0.9",
    "pulumi-minio>=0.16.4",
    "omegaconf>=2.3.0",
    "slack>=0.0.2",
    "s3fs>=2025.10.0",
    "slack-sdk>=3.38.0",
 ]

[tool.setuptools.packages.find]
where = ["."]
include = ["clap*", "infrastructure*", "infrastructure/components*", "helper*", "data_collection*"]
